Add strings for localisation of Wikivoyage RelatedSites links
These strings are currently hard-coded in English in
InitialiseSettings.php, with no way to localise other than by Bugzilla
request.

This will have no effect until a change to the RelatedSites extension is
merged (Id6b1aac8e82) and also to InitialiseSettings (I4f353cfb2cc).
However, it is harmless if this change is merged before the other changes.

Bug: 68359
Change-Id: Icd91c80a2f3eace388074a6ad0117a5ecea00c6b