Add default Wikimedia privacy policy link
On new wikis and wikis which did not customise the message locally,
the footer points to a non-existing page and the site is effectively
lacking a privacy policy.

The core function Skin:privacyLink() only accepts a title, need to
use the interwiki. As first step, point to what most sites do:
https://wikimediafoundation.org/wiki/Privacy_policy

Change-Id: Id99272e2227d1ce9a5d4a229ccdb0c9c43532ef7