Wikimedia-specific error message for OAuth error E008
This error happens when a CentralAuth user account that's being
used for Auth authorization is not properly merged. Due to the
many ways CentralAuth can be configured, the default message is
vague and not very helpful. On Wikimedia wikis, this error should
only occur as a result of some internal error, most likely some
race condition, so a more helpful message can be used.

Bug: T105105
Change-Id: I7127e8adc0bac776c9fde59dc46ddff17f6b39ed