Add i18n messages for Meta Wiki, MediaWiki wiki and Wikispecies site links on Wikidata
Also updated the description of the wikibase-sitelink-sitename* messages
as they now have a different, more limited use.

As well, removed wikibase-sitelink-commons message as it is (99% sure) unused.
We used to have a "Wikimedia Commons" site link section when it was the
only "special" site but now we include it in the "Special" site links
section.

Change-Id: I62ea17294c37ebdf65270f696f846201032c54fe