Add "Identifiers" section heading for wikidata.org
This message is currently unused, but will be used in a few days while
we are working on T117421. When deploying the new feature it would be
good if the message is already translated in the most relevant languages.

This message is here and not in the Wikibase.git because it depends on
configuration and is not part of the Wikibase default settings.

Bug: T120648
Change-Id: I36bd02e66d2629976567590656f4eff9815f5dc6