Creates an override for WMF wikis for MediaWiki:Delete-toobig
Administrators from WMF wikis often do not know what to do with pages
with large edit histories in order to have them deleted. This override
for WMF projects will direct the administrators to the stewards, which
have the appropriate rights to perform such operation; something that
not all users know. Hopefully this will add further clarification.

Bug: T139797
Change-Id: I9ba7eddd5f720362f168d985941a34dd7c6bdbd7