Remove links to main page from interlanguage search messages
Per T151649, it would be better to use the bare wikilink to link to a wiki
so that we do not have to rely on redirects from [[Main Page]] on that
wiki to the appropriately localised page. e.g. en.wikipedia.org/wiki/fr:
does not rely on a redirect on the wiki that you're linking to, whereas
en.wikipedia.org/wiki/fr:Main_Page does.

Bug: T151649
Change-Id: I00e7bae64270ea09b39a853001ea4173568b7071