RCFilters Guided tour: Append the dot to the popup
The dot should be in a higher DOM level so it isn't being cut off
by the div it's in. Append the dot to the bottom of the popup and
then position it properly, to make sure that the animation and the
dot itself are on top of the popup header and aren't cut off.

Bug: T151006
Change-Id: I063c303442f948ddbc375bafbdeb6cf741a5ab5c