RCFilters guided tour: Attach dot back to button
Now that the highlight button is moving (and not 'sticky') the dot
being in the overlay above requires readjusting of positioing.

Instead, attach it back to the button - and since it's no longer
sticky, it should be possible to z-index it above. Cancel
overflow:hidden; so this works.

While this doesn't quite depend on it, it is related to
Iac3b8eabdeaee009

Bug: T162971
Change-Id: Iaaebf9f45ff1869e1d61b26dc8e3bdbfc2c6cf85