Update wikidata-copyright to match wd.o/MediaWiki:Wikidata-copyright
This message had been overridden on the MediaWiki:Wikidata-copyright
page on wikidata.org.
But this copy is not being translated, so all languages that are no EN
have the older message...

Move the up to date message back to WikimediaMessages so that it
can be translated on translate wiki.
This also prepares the message for updating prior to the WikibaseLexeme
deployment.

Once this change is merged and deployed the mediawiki page for the
message should be removed so that this message is actually used.

Bug: T169333
Change-Id: Icc88f3834fd4a7f946a7bbd4c7bf7c833fedf680