Use singular they in CC0 explanation
This is more inclusive than “he or she”, and also shorter.

It also introduces a difference between the English version of this
message and the original license deed on Creative Commons [1], which
were previously identical except for markup. However, in other languages
(I checked German) there are already significant differences between our
and Creative Commons’ translation of the deed / message, so I assume
this change is acceptable.

[1]: https://creativecommons.org/publicdomain/zero/1.0/deed.en

Bug: T223149
Change-Id: I9801a71be290dd230800b3fa8455eae357e3a4aa