Create an override for 'globalblocking-ipblocked-range' for Wikimedia
Creates a Wikimedia override for the 'globalblocking-ipblocked-range
message from the GlobalBlocking extension which is called
'wikimedia-globalblocking-ipblocked-range', similarly to what we've done
for other GlobalBlocking messages so Wikimedia users affected by global
IP blocks knows where to request more information or appeal them.

Documentation was tweaked a bit on `includes/WikimediaMessagesHooks.php`
so it's clear we're talking about global blocks, just in case.

Bug: T243974
Change-Id: If0cac4312a6ad0b6ce2b144cce9d65f563d556da