Restore terms of use link URL to wikis
The terms of use should be overriden by an update to the message
key to mobile-frontend-terms-url. It should not use a hook, as
that requires knowledge of Minerva internals which are not going to
be true very soon!

Nikerabbit points out that Special:MyLanguage is not available on
foundation.wikimedia.org so this is dropped.

Bug: T252375
Change-Id: I82e647b8a4c7e0d489976721bde7673153e774db