Fix missing license name in mobile editors on Wikidata and Commons
The 'MobileLicenseLink' hook handler was not correctly handling the
two mostly unrelated things that this hook does. Setting an empty
license name was okay for the page footer, because it's not used
there, but it was causing the license name to be missing in mobile
editors.

Fix the bug and add some code comments about it.

Bug: T284097
Change-Id: I70302c629d970521914023b5c5ce102fba62449f