Do not use raw HTML for 'wikimedia-mobile-license-links'
Including the 'title' attribute is not worth it, especially because
this message is used on mobile, where these tooltips are never shown.

Including the 'target="_blank"' attribute is the wrong approach,
it doesn't get applied when we're not overriding the message.
Instead MobileFrontend code should add it, and it already does!

Replacements done in all existing translations:

Find:
  "wikimedia-mobile-license-links": "<a href=\\"(.+?)\\" title=\\".+?\\" target=\\"_blank\\">(.+?)<\/a>(.+?)<a href=\\"(.+?)\\" title=\\".+?\\" target=\\"_blank\\">(.+?)<\/a>",
Replace with:
  "wikimedia-mobile-license-links": "[$1 $2]$3[$4 $5]",

Change-Id: I2a04969b2c3df896900f6775a510dfd727ef9a5a