Move licensing logic from MobileFrontend to WikimediaMessages
This puts all logic relating to licenses in the same place without
the code in MobileFrontend adding indirection.

This is a backwards compatible change. The only thing visible to a user
will be on desktop Minerva, the two licenses shown in the footer should
be the same. The duplicate license is removed in a follow up.

Bug: T289401
Change-Id: I3473908e08f643b4c60aafc470dd124516955323