Fix Wikinews license link in mobile editor
This hack is necessary because MobileFrontend doesn't always display
the license defined in the usual config settings (T296791). I've
proposed a patch to fix that, but it was rejected (Id918234bab).

We need two versions of the link, because huwikinews has a unique
license. (The wiki is closed, so probably no one will see this, but I
think we shouldn't allow it to be out of sync with the site license.)

Bug: T284097
Bug: T296791
Change-Id: Ifdf8583c55d45144d85d0f6aa5e43b74c39839d3