Introduce SiteAdminHelper to allow site admins to cherry pick styles
As flagged in T358078 the current approach in Minerva doesn't give enough
control to site admins.

This newly proposed module would allow site admins to cherry pick
the styles they want to maintain themselves and the styles they don't.

Notes:
* CSS rules are direct copies of Minerva. If we need to adjust them
suggest limiting to follow up if it would cause visual regression. When
we enable this it should be possible to verify in Pixel there is no
change.
* Nothing is disabled by default
* The new module is intentionally not yet enabled on a skin. You will
need to test it using w/load.php?lang=en&modules=ext.wikimediamessages.styles
* content is replaced with mw-parser-output
* Some rules have been limited to the skin--responsive class - this means
they won't apply on Vector 2022 until it becomes a responsive skin. This avoids
shipping styles to Vector 2022 which we are not yet ready for.
* Took opportunity to switch to design tokens instead of custom variables
and rename camel case variables to hyphenated.
* Cleaned up ambox rules - the issue-B class is a remnant of an A/B test
* Switches from CSS variables to LESS variables so other skins can
potentially benefit in future.

Bug: T360386
Change-Id: I3cd7ff398008b5505401340729c7f0a3d962b2c0