Update translations of Wikifunctions "copyrightwarning" messages
Migrate translations of the raw HTML messages removed in Ic7be4ba721
to the wikitext messages added in If68775e489. Doing this in a
separate commit for ease of review.

Omitted one translation where the HTML formatting was incorrect.

Done with find-and-replace in three steps:

1. Update first link
Find: (wikifunctions-edit.*)<a\b.*?\bhref=[\\'"]*([^\\'" ]+).*?>([^<]+)</a>
Replace with: $1[$2 $3]

2. Update second link
Find: (wikifunctions-edit.*)<a\b.*?\bhref=[\\'"]*([^\\'" ]+).*?>([^<]+)</a>
Replace with: $1[$2 $3]

3. Update message names
Find: wikifunctions-edit-copyrightwarning
Replace with: wikifunctions-editing-copyrightwarning

Bug: T375882
Change-Id: I61c7cd979a98b14a186fa7194ef20dc293f564eb