Move full stop outside of sentence-final link in copyright footers
In several copyright footer messages that end with a link to the
Privacy Policy, the full stop at the end of the sentence was placed
inside that link. It should probably be outside of it.

This mistake was present in the original raw HTML messages, and was
replicated to the wikitext ones in I225658de26. I've only noticed it
now; I guess it's easier to spot with wikitext markup.

If this is merged, I will make the same corrections in translations
on translatewiki.net that repeat the same mistake.

Change-Id: I1ff06c7abeee149551bcd2de543318e1f803afba