Update the English name of mni.wikipedia
The spelling "Meitei" or "Meetei" is inconsistent,
and there are several more ways to write this name
in English.

The language has another name in English: "Manipuri".
Its spelling is consistent, and it's probably a more common
name, too: it's used in ISO 639, CLDR, the Manipur Gazette,
which publishes the laws of the State of Manipur in India,
and Encyclopedia Britannica.

Bug: T343113
Change-Id: Ieea24e73438bc87603a7359bcd8451e3ccc40a83