Use ccorg interwiki for CC license text links
Most of these links already use the `creativecommons` interwiki; the new
`ccorg` one handles the remaining links that it couldn't.

wm-license-cc-pd-mark-link is left alone because it's just a link and
I'm not sure if it can be adapted to wikilink syntax.

The goal of this is to reduce the number of external links on Commons.

Bug: T343131
Change-Id: If185f31ca9d87d7dd978d32896dc61fbc8f0c3ed