Add new wm-license-cc-pd-mark-wikilink message
This is the interwiki version of wm-license-cc-pd-mark-link. Once
deployed, we'll update Commons' Template:Cc-pd-mark-footer to use the
new message and a wikilink instead of an external one and then the old
link message will be removed.

Because this is just a transformation of the existing message, I
automatically converted all the existing translations so it doesn't need
to be done manually again.

Bug: T343131
Change-Id: Iba08ff1c888a891de956c174c668012ebcab1697