Stop translating sitesupport-url
Don't ask translators to copy a URL, changing a uselang parameter. This
can be accomplished in software instead, although there are some
limitations as noted on the linked task.

The translations here were generally unused anyway.

Set the campaign to the domain name, for consistency with the message
override in addWiki.php.

Switch from utm to wmf prefix to avoid browser link stripping.

Bug: T367361
Bug: T379205
Change-Id: I83ce7b6c6cdf3de4a193ae9712471823f53435e6