Sort regions and countries with Collator
In some languages, country names have accent and php by default
puts the accented characters after the non-accented charcters.
For instance: A, B, ..., Z, Å
This is unnatural, the As should be grouped together regardless
of accents.

To fix this, this patch proposes to use the Collator PHP class.
It accepts a locale code and performs locale-aware string
comparisons.

Bug: T391230
Change-Id: I2fc7f7214e22b771160de030efb40560ec625724