ArticleCountryFiltersRegistry: handle invalid language code
Some invalid language codes cause the Collator constructor
to throw a 'Constructor failed' error.

This patch tries to get a collator instance using the
Collator::create factory function instead of the
constructor since it returns null instead of throwing
an error. In which case, the "root" collator is used
as a decent fallback.

Bug: T400710
Change-Id: I6fce84ae4d6d659ddbff70ccb853a1ad15c877ea