tests: skip test when cldr is not available
On CI, we would like to remove the recursive injection of dependencies.
WikimediaMessages has an integration test with cldr which is otherwise
not needed by extensions depending on WikimediaMessages.

Skip the cldr integration test when it is not loaded. This will save us
from having WikimediaMessages added on CI for any extension loading
CirrusSearch.

Bug: T389998
Change-Id: Idef25dd83209f0f00ccab1baed6bb81956204874